--- src/test/java/com/revature/runner/MainRunner.java
@@ -1,23 +1,0 @@
-package com.revature.runner;
-
-import com.revature.pages.*;
-import io.cucumber.testng.AbstractTestNGCucumberTests;
-import io.cucumber.testng.CucumberOptions;
-import io.github.bonigarcia.wdm.WebDriverManager;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.support.ui.WebDriverWait;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
-
-import java.time.Duration;
-import java.util.HashMap;
-import java.util.Map;
-
-
-@CucumberOptions(
-        features="src/test/resources/features",
-        glue="com.revature.steps",
-        tags = ""
-)
-public class MainRunner extends AbstractTestNGCucumberTests {}
