+++ src/test/java/com/revature/runners/DarkModeRunner.java
@@ -1,4 +1,4 @@
-package com.revature.runner;
+package com.revature.runners;
 
 import io.cucumber.testng.AbstractTestNGCucumberTests;
 import io.cucumber.testng.CucumberOptions;
@@ -7,6 +7,6 @@ import io.cucumber.testng.CucumberOptions;
 @CucumberOptions(
         features="src/test/resources/features",
         glue="com.revature.steps",
-        tags = "@orders"
+        tags = "@dark-mode"
 )
-public class OrdersRunner extends AbstractTestNGCucumberTests {}
+public class DarkModeRunner extends AbstractTestNGCucumberTests {}
