+++ src/test/java/com/revature/runners/NotificationRunner.java
@@ -1,4 +1,4 @@
-package com.revature.runner;
+package com.revature.runners;
 
 import io.cucumber.testng.AbstractTestNGCucumberTests;
 import io.cucumber.testng.CucumberOptions;
@@ -7,6 +7,6 @@ import io.cucumber.testng.CucumberOptions;
 @CucumberOptions(
         features="src/test/resources/features",
         glue="com.revature.steps",
-        tags = "@cart"
+        tags = "@notification"
 )
-public class CartRunner extends AbstractTestNGCucumberTests {}
+public class NotificationRunner extends AbstractTestNGCucumberTests {}
