+++ src/test/java/com/revature/runners/OrdersRunner.java
@@ -1,4 +1,4 @@
-package com.revature.runner;
+package com.revature.runners;
 
 import io.cucumber.testng.AbstractTestNGCucumberTests;
 import io.cucumber.testng.CucumberOptions;
@@ -7,6 +7,6 @@ import io.cucumber.testng.CucumberOptions;
 @CucumberOptions(
         features="src/test/resources/features",
         glue="com.revature.steps",
-        tags = "@search"
+        tags = "@orders"
 )
-public class SearchRunner extends AbstractTestNGCucumberTests {}
+public class OrdersRunner extends AbstractTestNGCucumberTests {}
